requirements.txt
requests
python-dotenv
python-dateutil
google-genai